Feat: create follower story component
